backend/requirements.txt
# ATRA Web IDE - Backend Dependencies

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.26.0
aiohttp>=3.9.0

# Database
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# SSE
sse-starlette>=1.8.0

# Utils
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Markdown
markdown>=3.5.0

# Фаза 3: кэш планов
cachetools>=5.3.0
redis>=5.0.0

# День 5: Prometheus метрики
prometheus-client==0.20.0
